Check permissions in endpoints and UI for projects. Refactor schema and permission checks.

+++ layers/projects/server/api/projects/[id].delete.ts
@@ -1,5 +1,8 @@
 // endpoint for deleting a project
 import { ProjectRepository } from "../../repository/project";
+import { assertHasPermissionOrThrow } from "#layers/shared/server/utils";
+import { ENTITY } from "#layers/projects/utils/constants";
+import { PERMISSIONS } from "#layers/shared/utils/permissions";
 
 export default defineEventHandler(async (event) => {
   const { user } = await requireUserSession(event);
@@ -11,6 +14,14 @@ export default defineEventHandler(async (event) => {
     });
   }
   const projectRepository = new ProjectRepository();
+
+  // Permission check
+  assertHasPermissionOrThrow(
+    user?.permissions,
+    ENTITY,
+    PERMISSIONS.PROJECTS_WRITE,
+    "You do not have permission to delete projects."
+  );
 
   return projectRepository.delete(id);
 });
